switch from corfu to company for in-buffer completion
Despite trying the `cape-capf-buster` it still seems like corfu for some reason
makes eglot/LSP end up in a unsynchronized state quite often, where the only
remedy is to revert the buffer or restart eglot. Something that is becoming
increasingly tedious. Several eglot discussions seems to support the theory of
corfu being the culprit for getting the LSP server out of sync ([1] [2]).

company-mode appears much more stable in this regard and seems to be preferred
(even targeted) by the eglot developer.

[1] https://github.com/joaotavora/eglot/discussions/1127
[2] https://github.com/joaotavora/eglot/discussions/1311#discussioncomment-7365719